move all newly created files to service
