--- spec/fixtures/data_export/my_anime_list/authorization_check.json
@@ -1,3 +1,0 @@
-{
-  "authorized": "OK"
-}
