+++ spec/fixtures/my_anime_list/sync/anime/blood-lad.json
