+++ spec/fixtures/my_anime_list/sync/anime/boku-no-pico.json
