+++ spec/fixtures/my_anime_list/sync/anime/toy-anime-list.json
